core-refactor: rename Element class to ElementBlock
