+++ packages/core/src/dom/element/element.ts
@@ -1,4 +1,4 @@
-import { Element } from './elementBlock.ts'
+import { ElementBlock } from '@block/element.ts'
 import { applyPropsToElement } from './property.ts'
 import { ElementProps, SvgProps } from './type.ts'
 
@@ -6,7 +6,7 @@ export const createElement = <TagName extends keyof HTMLElementTagNameMap>(
   tagName: TagName,
   props: Partial<ElementProps<TagName>> = {},
 ) => {
-  const block = new Element({ element: document.createElement(tagName) })
+  const block = new ElementBlock({ element: document.createElement(tagName) })
   applyPropsToElement(block, props)
   return block
 }
@@ -16,7 +16,7 @@ export const createSvgElement = (
   props: Partial<SvgProps> = {},
 ) => {
   // @ts-ignore
-  const block = new Element({ element: svgElement.cloneNode(true) })
+  const block = new ElementBlock({ element: svgElement.cloneNode(true) })
   applyPropsToElement(block, props)
   return block
 }
