+++ packages/core/src/dom/element/util.ts
@@ -1,9 +1,9 @@
-import { Element } from './elementBlock.ts'
+import { ElementBlock } from '@block/element.ts'
 import { applyPropsToElement } from './property.ts'
 import { AllElementProps } from './type.ts'
 
 export const overrideElement = (
-  block: Element,
+  block: ElementBlock,
   props: Partial<AllElementProps>,
 ) => {
   applyPropsToElement(block, props)
