+++ packages/core/src/dom/index.ts
@@ -1,13 +1,13 @@
-export { For } from '@children/for.ts'
-export { Switch } from '@children/switch.ts'
-export { Toggle } from '@children/toggle.ts'
-export { Suspense } from '@children/suspense.ts'
+export { For } from '@flow/for.ts'
+export { Switch } from '@flow/switch.ts'
+export { Toggle } from '@flow/toggle.ts'
+export { Suspense } from '@flow/suspense.ts'
 export { component } from '@component/component.ts'
 export { root } from '@component/root.ts'
 export * from '@element/elementMap.ts'
 export { overrideElement } from '@element/util.ts'
-export { Component } from '@component/componentBlock.ts'
-export { Element } from '@element/elementBlock.ts'
+export { ComponentBlock } from '@block/component.ts'
+export { ElementBlock } from '@block/element.ts'
 export { isTextNode } from '@type/guard.ts'
 export { isRvjsFunction, isRvjsObject } from '@type/guard.ts'
 export type { ComponentFn } from '@component/component.ts'
